mover a cyclope.utils mixin para loguearse en tests funcionales
